+++ src/css/NavbarResponsive.css
@@ -1,22 +1,326 @@
 @media only screen and (max-width: 1200px) {
-    .header_logo img {
-      width: 110px;
-      height: 40px;
-      margin-top: 20px;
-    }
-    #lang {
-      left: 120px;
-      font-size: 14px;
-    }
-    .icon-container {
-      width: 14%;
-      bottom: 130px;
-    }
-    .sun,
-    .moon {
-      padding: 12px;
-    }
+  .header{
+    display: flex;
+    flex-direction: column;
+}
+.header-first{
+    display: flex;
+    justify-content: space-between;
+    gap: 50px;
+}
+.header_logo img {
+    width: auto;
+    height: auto;
+    margin-top: 30px;
+    max-width: 100%;
+}
+#lang {
+    position: absolute;
+    top: 35px;
+    left: 150px;
+    font-size: 15px;
+    cursor: pointer;
+    z-index: 1000;
+}
+.header__menu {
+    list-style: none;
+    padding: 0;
+    margin: 0;
+    /* display: flex;
+    flex-direction: column; */
+    gap: 10px;
+}
+.header__menu li {
+    color: black;
+    margin-top: 20px;
+    padding: 20px;
+    text-decoration: none;
   }
+.header__menu__item2.active {
+    background-color: #04AA6D;
+    color: white;
+}
+.header__menu__item2:hover:not(.active) {
+    background-color: blue;
+    color: white;
+    cursor: pointer;
+}
+.header__menu__item1.active {
+    background-color: #04AA6D;
+    color: white;
+}
+.header__menu__item1:hover:not(.active) {
+    background-color: blue;
+    color: white;
+    cursor: pointer;
+}
+.header__menu__item{
+    font-size: 20px;
+}
+.icon-container {
+    width: auto;
+    position: absolute;
+    bottom: 150px;
+    left: 30px;
+    display: flex;
+    gap: 5px;
+    background-color: gray;
+    align-items: center;
+    padding: 2px 3px;
+    cursor: pointer;
+    /* max-width: 100%; */
+    border-radius: 10%;
+}
+.sun {
+    background-color: white;
+    color: gray;
+    padding: 5px;
+    border-radius: 10%;
+}
+.moon {
+    color: white;
+    margin: auto;
+}
+
+
+body.dark-mode {
+  background-color: #333; 
+  color: #fff;
+}
+
+.sidebar.dark-mode {
+  background-color: #333; 
+  color: #fff;
+}
+
+.sidebar.dark-mode .header__menu li,
+.sidebar.dark-mode .footer_menu li {
+  color: #fff; 
+}
+
+.sidebar.dark-mode .icon-container .sun {
+  display: none;
+}
+
+.sidebar.dark-mode .icon-container .moon {
+  display: block;
+}
+.footer {
+    display: flex;
+    justify-content: space-between;
+    gap: 30px;
+    cursor: pointer;
+    position: absolute;
+    bottom: 90px;
+    left: 0px;
+    flex-wrap: wrap;
+}
+.footer_menu {
+    list-style: none;
+    display: flex;
+    gap: 20px;
+}
+.footer_menu_item {
+    cursor: pointer;
+    position: relative;
+}
+.footer_menu_item li {
+    text-decoration: none;
+}
+.footer_menu_item1, .footer_menu_item2, .footer_menu_item3 {
+    text-decoration: none;
+    cursor: pointer;
+    color: black;
+    position: relative;
+    overflow: hidden;
+}
+.footer_menu_item1::after, .footer_menu_item2::after, .footer_menu_item3::after {
+    content: '';
+    position: absolute;
+    left: 0;
+    bottom: -1px;
+    width: 100%;
+    height: 2px;
+    background-color: black;
+    transform: scaleX(0);
+    transform-origin: left;
+    transition: transform 0.5s ease;
+}
+.footer_menu_item1:hover::after, .footer_menu_item2:hover::after, .footer_menu_item3:hover::after {
+    animation: underlineDelay 0.5s ease 0.1s forwards;
+}
+@keyframes underlineDelay {
+    to {
+        transform: scaleX(1);
+    }
+}
+
+
+    
+  }
+
+
+  @media only screen and (max-width: 1098px) {
+    .header{
+      display: flex;
+      flex-direction: column;
+  }
+  .header-first{
+      display: flex;
+      justify-content: space-between;
+      gap: 50px;
+  }
+  .header_logo img {
+      width: auto;
+      height: auto;
+      margin-top: 30px;
+      max-width: 100%;
+  }
+  #lang {
+      position: absolute;
+      top: 35px;
+      left: 150px;
+      font-size: 15px;
+      cursor: pointer;
+      z-index: 1000;
+  }
+  .header__menu {
+      list-style: none;
+      padding: 0;
+      margin: 0;
+      /* display: flex;
+      flex-direction: column; */
+      gap: 10px;
+  }
+  .header__menu li {
+      color: black;
+      margin-top: 20px;
+      padding: 20px;
+      text-decoration: none;
+    }
+  .header__menu__item2.active {
+      background-color: #04AA6D;
+      color: white;
+  }
+  .header__menu__item2:hover:not(.active) {
+      background-color: blue;
+      color: white;
+      cursor: pointer;
+  }
+  .header__menu__item1.active {
+      background-color: #04AA6D;
+      color: white;
+  }
+  .header__menu__item1:hover:not(.active) {
+      background-color: blue;
+      color: white;
+      cursor: pointer;
+  }
+  .header__menu__item{
+      font-size: 20px;
+  }
+  .icon-container {
+      width: auto;
+      position: absolute;
+      bottom: 150px;
+      left: 30px;
+      display: flex;
+      justify-content: space-between;
+      background-color: gray;
+      align-items: center;
+      padding: 2px 3px;
+      cursor: pointer;
+      /* max-width: 100%; */
+      border-radius: 10%;
+  }
+  .sun {
+      background-color: white;
+      color: gray;
+      padding: 9px;
+      border-radius: 10%;
+  }
+  .moon {
+      color: white;
+      margin: auto;
+  }
+  
+  
+  body.dark-mode {
+    background-color: #333; 
+    color: #fff;
+  }
+  
+  .sidebar.dark-mode {
+    background-color: #333; 
+    color: #fff;
+  }
+  
+  .sidebar.dark-mode .header__menu li,
+  .sidebar.dark-mode .footer_menu li {
+    color: #fff; 
+  }
+  
+  .sidebar.dark-mode .icon-container .sun {
+    display: none;
+  }
+  
+  .sidebar.dark-mode .icon-container .moon {
+    display: block;
+  }
+  .footer {
+      display: flex;
+      justify-content: space-between;
+      gap: 30px;
+      cursor: pointer;
+      position: absolute;
+      bottom: 90px;
+      left: 0px;
+      flex-wrap: wrap;
+  }
+  .footer_menu {
+      list-style: none;
+      display: flex;
+      gap: 20px;
+  }
+  .footer_menu_item {
+      cursor: pointer;
+      position: relative;
+  }
+  .footer_menu_item li {
+      text-decoration: none;
+  }
+  .footer_menu_item1, .footer_menu_item2, .footer_menu_item3 {
+      text-decoration: none;
+      cursor: pointer;
+      color: black;
+      position: relative;
+      overflow: hidden;
+  }
+  .footer_menu_item1::after, .footer_menu_item2::after, .footer_menu_item3::after {
+      content: '';
+      position: absolute;
+      left: 0;
+      bottom: -1px;
+      width: 100%;
+      height: 2px;
+      background-color: black;
+      transform: scaleX(0);
+      transform-origin: left;
+      transition: transform 0.5s ease;
+  }
+  .footer_menu_item1:hover::after, .footer_menu_item2:hover::after, .footer_menu_item3:hover::after {
+      animation: underlineDelay 0.5s ease 0.1s forwards;
+  }
+  @keyframes underlineDelay {
+      to {
+          transform: scaleX(1);
+      }
+  } 
+}
+
+
+
+
   @media only screen and (max-width: 992px) {
     .header_logo img {
       width: 100px;
